Move send/receive buffers to their own modules
